+++ tomcat/src/main/java/org/apache/catalina/servlet/RequestMapper.java
@@ -7,17 +7,17 @@ import org.apache.catalina.http.HttpRequest;
 
 public class RequestMapper {
 
-    private static final Set<Controller> controllers = Set.of(
+    private static final Set<AbstractController> controllers = Set.of(
             new LoginController(),
             new RegisterController()
     );
 
-    private static final Controller resourceController = new ResourceController();
+    private static final AbstractController resourceController = new ResourceController();
 
     private RequestMapper() {
     }
 
-    public static Controller getController(HttpRequest request) {
+    public static AbstractController getController(HttpRequest request) {
         return controllers.stream()
                 .filter(controller ->
                         request.URIStartsWith(controller.getClass().getAnnotation(RequestMapping.class).value())
